+++ crates/github_scbot_database/src/models/adapter.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use super::{
     auth::{
         AccountDbAdapter, DummyAccountDbAdapter, DummyExternalAccountDbAdapter,
@@ -49,7 +47,7 @@ pub struct DatabaseAdapter {
 
 impl DatabaseAdapter {
     /// Creates a new database adapter.
-    pub fn new(pool: Arc<DbPool>) -> Self {
+    pub fn new(pool: DbPool) -> Self {
         Self {
             account_adapter: AccountDbAdapter::new(pool.clone()),
             external_account_adapter: ExternalAccountDbAdapter::new(pool.clone()),