+++ crates/github_scbot_types/src/issues.rs
@@ -143,7 +143,7 @@ pub struct GhIssueCommentEvent {
     /// Repository.
     pub repository: GhRepository,
     /// Organization.
-    pub organization: GhUser,
+    pub organization: Option<GhUser>,
     /// Sender.
     pub sender: GhUser,
 }