+++ CHANGELOG.md
@@ -7,16 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2021-12-02
+
 ## Added
 
+- Handle conflict status on pull requests.
 - Creation of a `github-scbot-sentry` crate for easy maintenance.
 - New `/debug` route (enabled with the `BOT_TEST_DEBUG_MODE` environment variable) to try error reporting.
 
 ## Changed
 
-- Use the `rsa` crate instead of `openssl` (to generate RSA keys)
+- Use the `rsa` crate instead of `openssl` (to generate RSA keys).
 - Reuse `just` instead of an adhoc crate.
-- Update Sentry to `0.23`, with adaptations in `sentry-actix` and `sentry-eyre`
+- Update Sentry to `0.23`, with adaptations in `sentry-actix` and `sentry-eyre`.
 
 ## [0.16.0] - 2021-10-27
 
@@ -178,7 +181,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Give rights to external accounts on specific repositories
 - Simple terminal UI interface to have an overview on pull requests
 
-[Unreleased]: https://github.com/sharingcloud/github-scbot/compare/v0.16.0...HEAD
+[Unreleased]: https://github.com/sharingcloud/github-scbot/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/sharingcloud/github-scbot/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/sharingcloud/github-scbot/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/sharingcloud/github-scbot/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/sharingcloud/github-scbot/compare/v0.13.1...v0.14.0